Refactored transliteration into new file, for future transliteration testing (pinyin, buckwalter?)
